+++ spec/rubocop/cop/lint/non_local_exit_from_iterator_spec.rb
@@ -90,6 +90,28 @@ describe RuboCop::Cop::Lint::NonLocalExitFromIterator do
       it { expect(cop.offenses).to be_empty }
     end
 
+    context 'when lambda is inside of block followed by method chain' do
+      let(:source) { <<-END }
+        RSpec.configure do |config|
+          # some configuration
+
+          if Gem.loaded_specs["paper_trail"].version < Gem::Version.new("4.0.0")
+            current_behavior = ActiveSupport::Deprecation.behavior
+            ActiveSupport::Deprecation.behavior = lambda do |message, callstack|
+              return if message =~ /foobar/
+              Array.wrap(current_behavior).each do |behavior|
+                behavior.call(message, callstack)
+              end
+            end
+
+            # more configuration
+          end
+        end
+      END
+
+      it { expect(cop.offenses).to be_empty }
+    end
+
     context 'when block in middle of nest is followed by method chain' do
       let(:source) { <<-END }
         transaction do